+++ src/AdminBar.php
@@ -10,6 +10,7 @@
 
 namespace wb\adminbar;
 
+use wb\adminbar\twigextensions\AdminBarTwigExtension;
 use wb\adminbar\variables\AdminBarVariable;
 use wb\adminbar\models\Settings;
 
@@ -65,6 +66,9 @@ class AdminBar extends Plugin
         parent::init();
         self::$plugin = $this;
 
+        // Add in our Twig extensions
+        Craft::$app->view->twig->addExtension(new AdminBarTwigExtension());
+
         // Register our site routes
         Event::on(
             UrlManager::className(),
@@ -83,24 +87,24 @@ class AdminBar extends Plugin
             }
         );
 
-/**
- * Logging in Craft involves using one of the following methods:
- *
- * Craft::trace(): record a message to trace how a piece of code runs. This is mainly for development use.
- * Craft::info(): record a message that conveys some useful information.
- * Craft::warning(): record a warning message that indicates something unexpected has happened.
- * Craft::error(): record a fatal error that should be investigated as soon as possible.
- *
- * Unless `devMode` is on, only Craft::warning() & Craft::error() will log to `craft/storage/logs/web.log`
- *
- * It's recommended that you pass in the magic constant `__METHOD__` as the second parameter, which sets
- * the category to the method (prefixed with the fully qualified class name) where the constant appears.
- *
- * To enable the Yii debug toolbar, go to your user account in the AdminCP and check the
- * [] Show the debug toolbar on the front end & [] Show the debug toolbar on the Control Panel
- *
- * http://www.yiiframework.com/doc-2.0/guide-runtime-logging.html
- */
+        /**
+         * Logging in Craft involves using one of the following methods:
+         *
+         * Craft::trace(): record a message to trace how a piece of code runs. This is mainly for development use.
+         * Craft::info(): record a message that conveys some useful information.
+         * Craft::warning(): record a warning message that indicates something unexpected has happened.
+         * Craft::error(): record a fatal error that should be investigated as soon as possible.
+         *
+         * Unless `devMode` is on, only Craft::warning() & Craft::error() will log to `craft/storage/logs/web.log`
+         *
+         * It's recommended that you pass in the magic constant `__METHOD__` as the second parameter, which sets
+         * the category to the method (prefixed with the fully qualified class name) where the constant appears.
+         *
+         * To enable the Yii debug toolbar, go to your user account in the AdminCP and check the
+         * [] Show the debug toolbar on the front end & [] Show the debug toolbar on the Control Panel
+         *
+         * http://www.yiiframework.com/doc-2.0/guide-runtime-logging.html
+         */
         Craft::info('AdminBar ' . Craft::t('adminBar', 'plugin loaded'), __METHOD__);
     }
 
@@ -124,7 +128,7 @@ class AdminBar extends Plugin
      *
      * @return bool Whether the plugin should be installed
      */
-    protected function beforeInstall(): bool
+    protected function beforeInstall():bool
     {
         return true;
     }
@@ -141,7 +145,7 @@ class AdminBar extends Plugin
      *
      * @return bool Whether the plugin should be updated
      */
-    protected function beforeUpdate(): bool
+    protected function beforeUpdate():bool
     {
         return true;
     }
@@ -158,7 +162,7 @@ class AdminBar extends Plugin
      *
      * @return bool Whether the plugin should be installed
      */
-    protected function beforeUninstall(): bool
+    protected function beforeUninstall():bool
     {
         return true;
     }
@@ -186,14 +190,24 @@ class AdminBar extends Plugin
      *
      * @return string The rendered settings HTML
      */
-    protected function settingsHtml(): string
+    protected function settingsHtml():string
     {
+        AdminBar::$plugin->bar->clearAdminBarCache();
+
+        $settings = $this->getSettings();
+
+        if (empty($settings->customLinks)) {
+            $settings['customLinks'] = [['','',0]];
+            //Craft::dd($settings['customLinks']);
+        }
+        //Craft::dd($settings);
+
         return Craft::$app->view->renderTemplate(
             'adminbar'
             . DIRECTORY_SEPARATOR
             . 'settings',
             [
-                'settings' => $this->getSettings()
+                'settings' => $settings
             ]
         );
     }